// lib/types/api.ts
export interface Zigbee2MQTTAPI {
    "bridge/logging": {
        message: string;
        level: LogLevel;
        namespace: string;
    };

    "bridge/state": {
        state: "online" | "offline";
    };
// ...
}